+++ marshmallow_jsonapi/__init__.py
@@ -1,4 +1,4 @@
 from .schema import Schema, SchemaOpts
 
-__version__ = "0.21.2"
+__version__ = "0.22.0"
 __all__ = ("Schema", "SchemaOpts")